Create a class GamePlayers including [Player] and Dealer
